feat: update session code according to DMTF 2025.3
