+++ redfish/openapi/dmtf/SessionService.v1_1_9.yaml
@@ -2,7 +2,8 @@ components:
   schemas:
     SessionService_v1_1_9_SessionService:
       additionalProperties: false
-      description: The `SessionService` schema describes the session service and its properties, with links to the actual list of sessions.
+      description: The `SessionService` schema describes the session service and its
+        properties, with links to the actual list of sessions.
       properties:
         '@odata.context':
           $ref: http://redfish.dmtf.org/schemas/v1/odata-v4.yaml#/components/schemas/odata-v4_context
@@ -27,38 +28,55 @@ components:
         Oem:
           $ref: http://redfish.dmtf.org/schemas/v1/Resource.yaml#/components/schemas/Resource_Oem
           description: The OEM extension property.
-          x-longDescription: This property shall contain the OEM extensions. All values for properties that this object contains shall conform to the Redfish Specification-described requirements.
+          x-longDescription: This property shall contain the OEM extensions. All
+            values for properties that this object contains shall conform to the Redfish
+            Specification-described requirements.
         ServiceEnabled:
-          description: An indication of whether this service is enabled. If `true`, this service is enabled. If `false`, it is disabled, and new sessions cannot be created, old sessions cannot be deleted, and established sessions can continue operating.
+          description: An indication of whether this service is enabled. If `true`,
+            this service is enabled. If `false`, it is disabled, and new sessions
+            cannot be created, old sessions cannot be deleted, and established sessions
+            can continue operating.
           nullable: true
           readOnly: false
           type: boolean
-          x-longDescription: This property shall indicate whether this service is enabled. If `true`, this service is enabled. If `false`, it is disabled, and new sessions shall not be created, old sessions shall not be deleted, and established sessions can continue operating.
+          x-longDescription: This property shall indicate whether this service is
+            enabled. If `true`, this service is enabled. If `false`, it is disabled,
+            and new sessions shall not be created, old sessions shall not be deleted,
+            and established sessions can continue operating.
         SessionTimeout:
-          description: The number of seconds of inactivity that a session can have before the session service closes the session due to inactivity.
+          description: The number of seconds of inactivity that a session can have
+            before the session service closes the session due to inactivity.
           format: int64
           maximum: 86400
           minimum: 30
           readOnly: false
           type: integer
-          x-longDescription: This property shall contain the threshold of time in seconds between requests on a specific session at which point the session service shall close the session due to inactivity. The session service shall support any value between the schema-specified minimum and maximum terms.
+          x-longDescription: This property shall contain the threshold of time in
+            seconds between requests on a specific session at which point the session
+            service shall close the session due to inactivity. The session service
+            shall support any value between the schema-specified minimum and maximum
+            terms.
           x-units: s
         Sessions:
           $ref: http://redfish.dmtf.org/schemas/v1/odata-v4.yaml#/components/schemas/odata-v4_idRef
           description: The link to a collection of sessions.
           readOnly: true
-          x-longDescription: This property shall contain a link to a resource collection of type `SessionCollection`.
+          x-longDescription: This property shall contain a link to a resource collection
+            of type `SessionCollection`.
         Status:
           $ref: http://redfish.dmtf.org/schemas/v1/Resource.yaml#/components/schemas/Resource_Status
-          description: The status and health of the resource and its subordinate or dependent resources.
-          x-longDescription: This property shall contain any status or health properties of the resource.
+          description: The status and health of the resource and its subordinate or
+            dependent resources.
+          x-longDescription: This property shall contain any status or health properties
+            of the resource.
       required:
         - '@odata.id'
         - '@odata.type'
         - Id
         - Name
       type: object
-      x-longDescription: This resource contains the session service properties for a Redfish implementation.
+      x-longDescription: This resource contains the session service properties for
+        a Redfish implementation.
       x-patternProperties:
         ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
           description: This property shall specify a valid odata or Redfish property.
