+++ redfish/openapi/merged/redfish-openapi.yaml
@@ -2914,20 +2914,10 @@ components:
           x-longDescription: This property shall contain the password for this session.
             The value shall be `null` in responses.
         SessionType:
+          $ref: '#/components/schemas/Session_SessionTypes'
           description: The active session type.
-          enum:
-          - HostConsole
-          - ManagerConsole
-          - IPMI
-          - KVMIP
-          - OEM
-          - Redfish
-          - VirtualMedia
-          - WebUI
-          - OutboundConnection
           nullable: true
           readOnly: true
-          type: string
           x-longDescription: This property shall represent the type of session that
             is currently active.
           x-versionAdded: v1_2_0
@@ -2951,6 +2941,68 @@ components:
       x-requiredOnCreate:
       - UserName
       - Password
+    Session_SessionTypes:
+      enum:
+      - HostConsole
+      - ManagerConsole
+      - IPMI
+      - KVMIP
+      - OEM
+      - Redfish
+      - VirtualMedia
+      - WebUI
+      - OutboundConnection
+      type: string
+      x-enumDescriptions:
+        HostConsole: The host's console, which could be connected through Telnet,
+          SSH, or other protocol.
+        IPMI: Intelligent Platform Management Interface.
+        KVMIP: A Keyboard-Video-Mouse over IP session.
+        ManagerConsole: The manager's console, which could be connected through Telnet,
+          SSH, SM CLP, or other protocol.
+        OEM: OEM type. For OEM session types, see the OemSessionType property.
+        OutboundConnection: A Redfish Specification-defined outbound connection. See
+          the 'Outbound connections' clause of the Redfish Specification.
+        Redfish: A Redfish session.
+        VirtualMedia: Virtual media.
+        WebUI: A non-Redfish web user interface session, such as a graphical interface
+          or another web-based protocol.
+      x-enumLongDescriptions:
+        HostConsole: This value shall indicate the session is the host's console,
+          which could be connected through Telnet, SSH, or other protocol. If this
+          session is terminated or deleted, the service shall close the connection
+          for the respective host console session.
+        IPMI: This value shall indicate the session is an Intelligent Platform Management
+          Interface session. If this session is terminated or deleted, the service
+          shall close the connection for the respective IPMI session.
+        KVMIP: This value shall indicate the session is a Keyboard-Video-Mouse over
+          IP session. If this session is terminated or deleted, the service shall
+          close the connection for the respective KVMIP session.
+        ManagerConsole: This value shall indicate the session is the manager's console,
+          which could be connected through Telnet, SSH, SM CLP, or other protocol.
+          If this session is terminated or deleted, the service shall close the connection
+          for the respective manager console session.
+        OEM: This value shall indicate the session is an OEM-specific session and
+          is further described by the OemSessionType property.
+        OutboundConnection: This value shall indicate the session is an outbound connection
+          defined by the 'Outbound connections' clause of the Redfish Specification.
+          The `OutboundConnection` property inside the `Links` property shall contain
+          the link to the outbound connection configuration. If this session is terminated
+          or deleted, the service shall disable the associated `OutboundConnection`
+          resource.
+        Redfish: This value shall indicate the session is a Redfish session defined
+          by the 'Redfish session login authentication' clause of the Redfish Specification.
+          If this session is terminated or deleted, the service shall invalidate the
+          respective session token.
+        VirtualMedia: This value shall indicate the session is virtual media session.
+          If this session is terminated or deleted, the service shall close the connection
+          for the respective virtual media session and make the media inaccessible
+          to the host.
+        WebUI: This value shall indicate the session is a non-Redfish web user interface
+          session. If this session is terminated or deleted, the service shall invalidate
+          the respective session token.
+      x-enumVersionAdded:
+        OutboundConnection: v1_7_0
     odata-v4_context:
       description: The OData description of a payload.
       format: uri-reference
